requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
pyjwt==2.8.0
python-multipart==0.0.9
pydantic[email]==2.6.1
email-validator==2.1.0.post1
pytest==8.0.0
pytest-asyncio==0.23.4
httpx==0.26.0
pytest-playwright==0.4.4
playwright==1.41.2
